pyproject.toml
[project]
name = "int-per-scene"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.7.0",
    "einops>=0.8.1",
    "huggingface-hub>0.25",
    "kornia>=0.8.1",
    "llvmlite>=0.44.0",
    "numba>=0.61.2",
    "numpy<2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "prettytable>=3.16.0",
    "rembg[cpu]>=2.0.66",
    "scikit-image>=0.25.2",
    "scipy>=1.15.3",
    "tabulate>=0.9.0",
    "timm>=1.0.15",
    "torch>=2.5.0",
    "torchaudio>=2.7.1",
    "torchvision>=0.22.1",
    "tqdm>=4.67.1",
    "transformers>=4.52.4",
]
